requirements.txt
# ViewSense Edge Worker
# Core
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
pyyaml==6.0.1
psutil>=5.9.0

# Computer Vision / AI
numpy<2.0.0
opencv-python-headless==4.9.0.80
ultralytics>=8.2.0
supervision>=0.18.0
torch>=2.1.0,<2.6.0
torchvision>=0.16.0,<0.21.0

# Demographics (gender + age estimation)
insightface>=0.7.3
onnxruntime>=1.16.0
